tdf#169480 Enhancement to recency navigation
Enhancement request for special handling for situations when a user
has scrolled/moved the view so that the cursor is no longer visible.
To handle this, recency navigation should check if the cursor is in
the view and if not, then scroll/move the view to make the cursor
visible without any manipulation of the back/forward stack.

Change-Id: Iaaa2b09108649ed12183dcc75b9bb03a7915458c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194147
Tested-by: Jenkins
Reviewed-by: Vernon, Stuart Foote <vsfoote@libreoffice.org>
Reviewed-by: Jim Raykowski <raykowj@gmail.com>