uitest: Wait for idle when calling close_doc for a database doc
The UITest calls .uno:CloseDoc at the end of create_db_in_start_center.
This seems to end up triggering some race condition which makes any
tests using it occasionally fail. I think what happens is this:

If the document is modified when CloseDoc is called then instead of
immediately closing the document it posts an event to the event queue
which causes the main thread to start closing the document
asynchronously. In the meantime the UI test thread continues and
executes Desktop::terminate. Both the main thread and the UI test thread
end up calling XFrameImpl::setComponent simultaneously. The main thread
does this with the solar mutex held because the event handling code
locks the mutex during the entire process. The UI test thread doesn’t
hold it, but parts of XFrameImpl::setComponent require the lock so it
blocks straight away. The first part of XFrameImpl::setComponent sets
the controller to the new one and then starts disposing the old one. The
disposing implementation for OApplicationController temporarily releases
the solar mutex while it does the potentially time consuming process of
disconnecting from the database (see the SolarMutexReleaser in
ODriverDelegator::shutdownConnection). This causes the UI test thread to
unblock and continue executing XFrameImpl::setComponent. Presumably it
bypasses trying to dispose the controller because that will have already
been set to null by the main thread. It will therefore jump straight to
disposing the old component window. This invokes VCLXWindow::dispose
which ends up disposing the ODataView with the solar mutex locked. Once
it is finished and the main thread has finished disconnecting from the
database it will be able to reacquire the solar mutex and continue
invoking OApplicationController::disposing. The next step calls
attachFrame which in turn calls attachFrame on the ODataView. However
the data view has been disposed in the meantime which makes it call a
method on a null pointer because m_pAccel is now null. This will end up
throwing an uncaught std::system_error.

I tried working around this one problem by checking whether the
ODataView has already been disposed in ODataView::attachFrame. However
this then just triggers a different problem where Base tries to close
the database connection in one thread while
ODriverDelegator::flushConnections is running in another and it also
ends up throwing a std::system_error.

Seeing as there seems to be a nest of bugs in this area this patch just
makes it wait for CloseDoc to finish before continuing the rest of the
unit test. If we want to fix the problems later we could make some
further unit tests that explicitly execute CloseDoc without waiting.

Change-Id: Ie531ada3d740e9e777d2b73ee162d4818fb6c477
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194181
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>