idletask: Wait with a condition_variable when not on the main thread
Judging by the comment that was in waitUntilIdleDispatched, I think the
intention of this helper class is to wait until the main loop is idle
and that any event processing has actually completed. However, if this
is called outside of the main thread then that doesn’t work if the event
processing temporarily releases the solar mutex. In that case, the
thread that is waiting for the IdleTask flag would wake up and set the
flag itself even though the main thread hasn’t actually finished
executing its event handling code yet.

To fix that this patch makes it instead wait with a
std::condition_variable instead of Yield when not running on the main
thread so that we can be sure that the main thread will invoke the idle
task instead of the waiting thread.

Change-Id: Ib5c6efe31dfb5e1a5039b43702e23bf0104cd403
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194180
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>