[API CHANGE] Reintroduce com.sun.star.ui.dialogs.ColorPicker
Bring back the com.sun.star.ui.dialogs.ColorPicker service
previously dropped in

    commit 5ba0ccb2964da46d32ade0c87baf2d407cb52a8a
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Jun 16 08:01:55 2025 +0200

        [API CHANGE] Drop ColorPicker, AsynchronousColorPicker

    The unpublished com::sun::star::cui::ColorPicker and
    com::sun::star::cui::AsynchronousColorPicker UNO services
    were previously used to be able to use ColorPickerDialog
    (implemented in cui) from svtools.

    That's no longer the case since

        commit 0905c23676fe34a5346683ee46568e4d1c58806a
        Author: Michael Weghorn <m.weghorn@posteo.de>
        Date:   Fri Jun 13 18:27:09 2025 +0200

            Stop using UNO abstraction for color picker

    , so drop the now unused services and the ColorPicker
    class that implemented them.

While the service was unpublished and undocumented, it was
still used in practice by third party code, see for
example the macro in [1] and further discussion in Gerrit
change [2].

Therefore bring back the service, but with a new
implementation in vcl that no longer uses the ColorPickerDialog
implemented in cui directly, but instead uses the
weld::ColorChooserDialog abstraction introduced in

    commit dae9472b27914807ec08062e8293a091e90cffb6
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Jul 25 14:26:06 2025 +0200

        tdf#167669 weld: Introduce weld::ColorChooserDialog

which supports the use of the native platform/toolkit
color dialogs (currently implemented for GTK and Qt), see
tdf#167669.

Don't reintroduce evaluation of the "Mode" parameter
to set a vcl::ColorPickerMode because that seems unused
by third-party code according to a quick Google search.

This makes the macro from [1] work again, and results in
the native GTK and Qt color pickers being used for the
case of the gtk3/qt6 VCL plugins on Linux.

[1] https://forum.openoffice.org/en/forum/viewtopic.php?f=21&t=99268
[2] https://gerrit.libreoffice.org/c/core/+/186538/comments/3af0d3d8_8b843d4e

Change-Id: Ic0f2ce2b289f37bf1fb32192d0f3b3deb99121ce
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194172
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <stephan.bergmann@collabora.com>