sd: pptx: fix footer moving places for MSO
In a pptx file with a slide master that has multiple slide layouts, the
footer would loose properties when saved in LO and opened in MSO.
This is due to the master de-duplicating logic that in trying to emulate
pptx's natural way of dealing with master and layout slide it treats
multiple Impress masters as equivalent. This is needed for most export
functions, but when exporting an element from an Impress slide that
references an Impress master (like a footer), we need to differentiate
exactly which one of the equivalent masters it comes from to make the
indices match.

Change-Id: I815c3db54256d934b69108c7268f6d7642c33cbb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194190
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
(cherry picked from commit 0f21756af8e74b0dd3f25fe4234feb067430838b)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195225
Tested-by: Jenkins
Reviewed-by: Jaume Pujantell <jaume.pujantell@collabora.com>
(cherry picked from commit f2736a8407ed006741a0d1f55b525e39783b29f1)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195359
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>