tdf#167125 vcl a11y: Don't completely lose keyboard focus in VerticalTabControl
Similar to the issue for Calc's conditional formatting
dialog fixed by

    commit 752adf6cc9d869bbcf34fcf09c80da502398cb81
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Nov 4 19:35:47 2025 +0100

        tdf#169006 sc a11y: Keep keyboard focus in conditional formatting dlg

, keyboard focus could get lost completely when switching
between tabs in the VerticalTabControl by using the
keyboard.

Sample steps to reproduce:

* start Writer using the gen vcl plugin
* "Format" -> "Page Style"
* click on the "Footer" tab
* press Tab to move focus to the "Footer on" checkbox
* press PgDown key to switch toh the last tab

Without this change in place, there would be
no more way to switch focus to anything using the
keyboard any more, as focus is still in the no
longer visible tab.

To fix this, move focus to the tab bar when
the VerticalTabControl switches between tabs
due to a keyboard event on a control in the
tab that was propagated to the VerticalTabControl
ancestor because it wasn't handled earlier.

Change-Id: Iacea0135c5f3e04d542f8d534540651d50196405
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194249
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins