uitest/framework: Throw an exception when a signal is received
When a signal is received such as when the unit test is killed by the
kill-wrapper, the framework catches the signal and terminates the
connection to LibreOffice. However it would then continue executing the
test. This makes it difficult to diagnose problems when the kill-wrapper
is invoked. This patch makes it instead throw an exception so that the
build log will have a backtrace. It also changes the resetSignalHandler
method to set the signal handlers to the DEFAULT rather than IGNORE,
because ignoring all subsequent signals seems like an unusual thing to
do and there is no comment explaining it so it seems likely to be a
mistake.

Change-Id: I8d8283cccd4de1ce55b0a6cd947c1647d28f462f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194321
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins