Make the weird FT_Done_MM_Var() trick in vcl work also on Windows
For Windows we build freetype as a static library. But we must build
freetype with DLL_EXPORT defined so that symbols like FT_Done_MM_Var
get exported by the DLL into which the static freetype library is
linked. That is mergedlo.dll or vcllo.dll.

Then in vcl's dlFT_Done_MM_Var() we can't just call
osl_getAsciiFunctionSymbol() with a nullptr for the first "module"
parameter and expect it to look through all modules, like dlsym() does
on Unixes. On Windows, osl_getAsciiFunctionSymbol() calls
GetProcAddress(), and passing a NULL first parameter to tht is simply
invalid. We must look for FT_Done_MM_Var in mergedlo.dll, or if that
isn't loaded, vcllo.dll.

Possibly it would be simpler to just pretend we don't HAVE_DLAPI in
this case and call FT_Done_MM_Var() directly.

Without this commit, FT_Done_MM_Var will not be found, so the code
will just call free() instead of FT_Done_MM_Var(), and if built using
the debugging C library that will lead to a strange assertion failure
in the C library. With the non-debugging C library it might perhaps
lead to some odd bug, or just not do whatever proper deallocation of
resources or something it is that FT_Done_MM_Var() does.

Change-Id: Ia7b726cc7167ba22ea8f3263ad85f8cd4ac5c483
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194608
Reviewed-by: Thorsten Behrens <thorsten.behrens@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/197213
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@iki.fi>