Move ValueSetWithTextControl from svx to sw
This control is exclusively used in
sw::sidebar::PageSizeControl, so there's no need
to have it as an svx class defined in a global svx
header.

(And I think that using existing
weld::Widgets instead of having that custom
logic would be preferred, too, so this control
also shouldn't be more widely used in the
future. See also discussion in tdf#169637.)

Change-Id: I8b040072a975c227fad2280d9475a3bddfc3b4dc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194487
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Code-Style: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins