tdf#169541 DataProvider avoid corrupt backreference
If a document has defined more than one calcext:data-mapping element,
the import builds a vector DataResources in the ExternalDataMapper.
Error was, that not the current last item in the vector was refreshed
but always the first one.

If more than one DataResource exist, those before the last one have a
corrupted backreference to the data provider. I have not found where
this happens. Therefore, I now create in each refresh a new one. That
is the same as done in the first refresh. If someone finds where it
becomes corrupt it could be fixed differently. However, since the
error causes a crash, we should use this solution in the meantime.

Change-Id: If489c8e599bf44333c4936c608392c22d0f02d14
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194433
Tested-by: Jenkins
Code-Style: Regina Henschel <rb.henschel@t-online.de>
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194691