tdf#169806: Display baseline grid for page line-spacing.
1. Add a new option to the Grid tab page for displaying the baseline
grid.
* Add a new checkbox to LibreOffice Writer -> Grid tab page to
enable/disable displaying of the grid.
* Make the new option hidden by default and show it only for Writer
normal view (see SwGridTabPage class).
* Rename the tab page title from "Grid" to "Grids" to reflect the new
content of the tab page (options for two different types of grids).

2. Add a new configuration option for displaying the baseline grid.
* Extend Writer configuration with a new section for baseline the grid
option(s).
* Introduce a new ConfigItem for the new baseline grid related option
to allow saving it to and loading it from the user preferences. (see
SwBaselineGridConfig class).
* Handle the possible read-only state of the configuration option and
lock the related checkbox on the grid tab page, if necessary.

3. Compute the proper position of the baseline grid's horizontal lines.
* First, extract the ComputeRegister() method to calculate the proper
height/ascent values used for register-true rendering.
* Use this ComputeRegister() method to calculate the proper
height/ascent values both for the text formatting code and for the
drawing code of the baseline grid.

4. Draw the baseline grid.
*  Draw the grid based on the reference style used by page
line-spacing, if specified.
* If there is no page line-spacing set, use the "standard" reference
style (i.e. "Body text").
* Draw the baseline grid on top of other objects (e.g. shapes,
frames), so the user can verify the text alignment inside these objects
as well.
* Trigger a repaint of the whole page if the baseline grid is visible
and page line-spacing settings are changed.

5. Don't display the baseline grid when printing.
* Don't display the grid for print preview, for printing,
or for exporting to PDF.

Change-Id: I40c4cd4beeab6b1416365e9e728bfe1a7756b5c5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195350
Reviewed-by: Tamás Zolnai <zolnaitamas2000@gmail.com>
Tested-by: Jenkins