sc: fix changing of borders and invalidation after change
When changing the theme and the ThemeColorChanger was run it would
not change the borders if only top, left and bottom border lines
would be set and not the right border line. This would not work
because the wrong variable to check if we introduced any change
for the borders would be used (bChanged and not bLineChanged).

This fixes the problem and adds a test for borders and this
specific scenario.

It also fixes invalidation after changing the theme colors.

Change-Id: Ia0529f70cdd37d1007ec1171d3debbb0f8a58bb2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194451
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Code-Style: Tomaž Vajngerl <quikee@gmail.com>