tdf#169713 svx: paint from last to first, not first to last
In my example document, there were two vcl checkboxes
stacked on top of each other. One was ticked, and the other wasn't.
If I scrolled the document, then un-ticked was painted,
but if I clicked nearby, it was painted as ticked.

I don't know why this was implemented in two opposite ways.
I changed it to match how vcl PaintHelper::~PaintHelper does it.

From my code-reading, it sounds like the highest heaven is FirstChild
so that would be the last thing painted (over top of everything else).
Thus I think PaintHelper is the accurate code.

Change-Id: I0a41bd883977ff6a4f26f260f015545f84028247
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194729
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>