tdf#169656 Crash upon starting animated GIF
this is caused by freeing an object that is currently on the callstack.

I dont have a complete fix here, this will fix it 99% of the time.

regression from
  commit e6a8c40e5051363d5825bd09131b86cf8a2b659b
  Author: Noel Grandin <noel.grandin@collabora.co.uk>
  Date:   Tue Sep 30 14:54:37 2025 +0200
  tdf#116975 Laggy behavior when clicking between table cells in impress
but only because it made it more obvious, this has been a lurking problem.

Change-Id: I4f2f56a54a74f2a51e7ca05421d3db82de47f2ca
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194502
Tested-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins
Code-Style: Noel Grandin <noel.grandin@collabora.co.uk>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>