sd: do not export canvas slide to pptx
by decreasing the page count by 1. As the canvas page and its notes page
is always at the end, `PPTWriterBase::CreateSlide` and `PPTWriterBase::CreateNotes`
won't be called for them.

The same cannot be done for the master canvas page as its not
necessarily the last master page. Since that master page will be unused,
it is safe to export that.

Signed-off-by: Mohit Marathe <mohit.marathe@collabora.com>
Change-Id: Ie24f594d45c08a99548363cc1e20cfff977ce519
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194521
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>
Code-Style: Michael Stahl <michael.stahl@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/196706