update to officeotron-0.8.4.jar
This reveals a problem with our export of the "dateCompatibility"
attribute. We actually only want to write that out when targeting
Word 2007, otherwise we want to use the more normal "date1904"
attribute.

On top of the original commit, this cherry-pick needs to add
a bunch of calls to skipValidation() because we have validation
turned on for more stuff on master, and this new officeotron
seems to generate a lot more errors.

Change-Id: I972bc7298ecf5026a76d109535a91c0c2544d08a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194522
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Code-Style: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Noel Grandin <noel.grandin@collabora.co.uk>
(cherry picked from commit efebbf0107c57a2c3c73251085b90d7e6c06d3b0)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194705
Tested-by: Jenkins