cool#13574 sw lok: introduce a redline render mode view option
Redlines (when shown) are decorated with an underline for inserts and
with a strikethrough for deletions. There is no way to render just the
old or the new version, which would be handy for a 2-up render result
(old on the left hand side, new on the right hand side).

Now fully hiding inserts or deletions would require not sharing the same
layout and would be quite complex to implement, it's possible to simply
not render the unwanted content, typically leaving white holes in the
render result.

The newly introduced SwRedlineRenderMode as a view option does exactly
this: keep the default behavior unchanged, but allow not rendering
inserts or deletions as a view option.

This requires no change to the LOK API: ther was a paint "mode"
parameter already (to handle normal slides, slide masters, slide notes
in Impress), which can be reused here for Writer's redline needs.

Change-Id: I21aee667913b916c030d067bbe8bb03a5f6b3ee2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194676
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
Code-Style: Miklos Vajna <vmiklos@collabora.com>