related tdf#166724 sc GetRange: last twip still in cell
GetRange indicates which cells are found
inside the provided rectangular area.
The problem is that providing the BottomRight point of a cell
to GetRange was resulting in a different cell being "found".

I understand that the end of one cell is the beginning of another,
and that GetRange calculates in TWIPs (where 1 twip = 1.7 mm100)
so imprecision can affect a point at the very edge.

However, if we reduce BottomRight by 1 twip,
then certainly GetRange should find the original cell back.

These (erroneous) formulas already existed in 'Initial import'.

I don't care if this gets reverted.
I just added it because it makes my calculation cleaner.
The unit test highly suggests that this has been hiding bugs...

Change-Id: I06d1c362cd8f80913ee695825260dd5b31b29bfb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194347
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>