cool#13574 sw lok: improve non-standard redline render mode colors
Instead of completely omitting insertions and deletions when requested,
just de-saturate the original colors when the content is to be
semi-hidden.

Do this by undoing the changes to SwTextPainter::DrawTextLine() and
instead let SwModule::GetInsertAuthorAttr() /
SwModule::GetDeletedAuthorAttr() decide if the standard rendering for
that redline is OK or we want a custom color.

If we want something custom, then halve the saturation in
lcl_FillAuthorAttr() to paint it semi-hidden.

Also adapt the testcase, to check the saturation instead of the number
of painted strings.

Change-Id: I62362463bd1b6d6968a503c802e17a56a1aea1b1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194774
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>