tdf#136905 tdf#169710 Use TreeView in PageSizeControl
So far, the custom ValueSetWithTextControl was used
in the "Page Size" dropdown (used in the "Layout"
tab of the Tabbed interface in Writer) to show what
is effectively a table with two columns (page size
and dimensions).

Stop using that custom control, and instead use
a weld::TreeView for the job, which avoids having
to do manual rendering and ensures it takes into
account the style, etc.

Use automatic size for the columns and give the
tree view its preferred height, which makes sure
that all entries are visible without scrolling.

Enable the GtkTreeView properties
activate-on-single-click and hover-selection
to keep highlighting the entry that the mouse
hovers over and activating an entry when its
clicked once.

Change-Id: I21038e65c2bf0d7b931f38c49a7da9a596983f36
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194741
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>