tdf#152030 Allow selection deletion of protected table of contents
"Protect against manual changes" seems not the same as read-only when
it comes to removing table of contents. For example, .uno:RemoveTableOf
removes a table of content regardless of the "Protect against manual
changes" setting. This patch changes the current behavior of displaying
the read-only content write-protected info message to not displaying
the message and allowing table of content deletion/cutting
when the entire table of content is included in a selection.

Change-Id: I4a14f0b1a17afbaacf211d7ddbe14f613761ebc5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194753
Reviewed-by: Jim Raykowski <raykowj@gmail.com>
Tested-by: Jenkins