tdf#168120 sc: Switch to weld::TreeView in CellLineStyleControl
Similar to how

    commit 85dc059a3990b41f93be6e163c2c8d2138f9e0e9
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Nov 27 22:29:52 2025 +0100

        tdf#136905 tdf#169710 Use TreeView in PageSizeControl

did elswhere, also switch from using a (different)
custom control to using a weld::TreeView in the dropdown
to select from a list of border styles in the Calc sidebar.

The logic for creating the images for the single border
styles in CellLineStylePopup::CreateImage is based on
the existing logic in CellLineStyleValueSet::UserDraw
(which was used so far).

Indexes change because weld::TreeView indexes are
0-based while the ValueSet ones are 1-based
(and use 0 as a special value).

It will also no longer apply when switching from using
CellLineStyleValueSet to using a weld::TreeView in an
upcoming commit, because CellLineStyleValueSet uses
a 1-based index while weld::TreeView uses a 0-based one.

This reduces custom widget logic and custom drawing
and makes this more consistent with other widgets
used elsewhere.

This also addresses the issue of a font inconsistent
with what gets used elsewhere being used
previously, as mentioned in tdf#168120.

To trigger the dropdown:

* start Calc
* select a cell
* open the "Properties" deck in the sidebar
* in the "Cell Appearance" section, click on
  the "Specify the borders of the selected cell"
  control and select "Outer Border"
* click on the "Select the line style of your borders"
  control (second control from the top on
  the right hand side in the "Cell Appearance" section).

CellLineStyleValueSet is now unused and will be
dropped in an upcoming commit.

Change-Id: I05c1e5f8f3689af02b4337c65b4aaba481c67e86
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/194776
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>